refactor: move internal links files
