+++ src/internal-links/parse-internal-links.test.ts
@@ -1,5 +1,5 @@
 import { Transformer } from 'markmap-lib'
-import { parseInternalLinks } from './internal-links'
+import { parseInternalLinks } from './parse-internal-links'
 
 describe('markmap plugins: internal links', () => {
   const transformer = new Transformer()
